feat: extract the collections list into different component.
